Added VDIP_FileExists and checks in SNSL_ParseConfig as well as SNSL_ParseConfigHeader to ensure that a config is created before trying to read from one. Cleaned up a lot of code and renamed several to following naming conventions.
TODO: Create a default nodes file to make sure that we don't run into the same problems when trying to parse an unavailable file. The function has been prototyped and the only problem is that I'm not sure how to create a blank file at this point. Maybe some kind of special value will have to be inserted.

git-svn-id: svn+ssh://svn.ece.msstate.edu/data/www/htdocs/courses/design/2010/wiresoil/svn@41 db668ce9-9f08-4dc0-9949-e28e8fe2f565